+++ tests/testthat/test-ab_property.R
@@ -54,4 +54,8 @@ test_that("ab_property works", {
   expect_error(ab_property("amox", "invalid property"))
   expect_error(ab_name("amox", language = "INVALID"))
   expect_output(print(ab_name("amox", language = NULL)))
+  
+  expect_equal(ab_name("21066-6", language = NULL), "Ampicillin")
+  expect_equal(ab_loinc("ampicillin"),
+               c("21066-6", "3355-5", "33562-0", "33919-2", "43883-8", "43884-6", "87604-5"))
 })