+++ NEWS.md
@@ -1,8 +1,18 @@
-# AMR 0.9.0.9011
-## <small>Last updated: 15-Jan-2020</small>
+# AMR 0.9.0.9012
+## <small>Last updated: 26-Jan-2020</small>
+
+### New
+* Support for LOINC codes in the `antibiotics` data set. Use `ab_loinc()` to retrieve LOINC codes, or use LOINC code for input in any `ab_*` function:
+  ```r
+  ab_loinc("ampicillin")
+  #> [1] "21066-6" "3355-5"  "33562-0" "33919-2" "43883-8" "43884-6" "87604-5"
+  ab_name("21066-6")
+  #> [1] "Ampicillin"
+  ```
 
 ### Changes
 * Speed improvement for `as.mo()` (and consequently all `mo_*` functions that use `as.mo()` internally), especially for the *G. species* format (G for genus), like *E. coli* and *K penumoniae*
+* Better support for determination of *Salmonella* biovars
 * Input values for `as.disk()` limited to a maximum of 50 millimeters
 * Added a lifecycle state to every function, following [the lifecycle circle of the `tidyverse`](https://www.tidyverse.org/lifecycle)
 * For in `as.ab()`: support for drugs starting with "co-" like co-amoxiclav, co-trimoxazole, co-trimazine and co-trimazole (thanks to Peter Dutey)